+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - macros can be declared inside a begin block within a cond macro and used in the scope surrounding the cond macro
 - `arkscript --version` and `arkscript --help` now output ArkScript version with the commit hash
 - `void Value::toString(std::ostream&, VM&)` now becomes `std::string Value::toString(VM&)`
+- removed `Node::operator<<` to replace it with `Node::debugPrint`
 
 ### Removed
 - removed unused `NodeType::Closure`